Added 'output info' entry to mdoel settings
Using this entry, the user can now specify aliases for the entries of
the output vector. Therefore the display of the entries name in the data window
(if available), the display of its name in the plot window title and the display of the
according unit at the y-axis is now possible.