Fixed error in LinearStateFeedback Controller
This error was caused by subtracting the closest equilibrium `eq` from
the copy of the state vector `x`. Due tu explicitly casting `eq` into a
2D-Array (`np.atleast_2d(eq)`) numpy's broadcasting rules got triggered
which would broadcast the shapes 4x1 (x) and 1x4 (eq.T) to 4x4.
After matrix multiplication with the gain vector `k` (1x4) one would
come up with a resulting output `u` 1x4.

Since the controller output is propagated to the model input, this is
these are the weird `input_values` that occurred in the model equations.