Bugfix #3 in plot handling
If a new PlotWidget is created but the current results don't offer a
result that fits, no PlotItem is created. When later the data is present,
it is never shown due to the missing plot item.

To fix this issue, if no data is available, an empty PlotItem with
correct settings is created.